+++ e2e/crm7/provider/submit-claim.spec.js
@@ -12,9 +12,18 @@ import {
     ClaimDetailsPage,
     WorkItemPage
 } from '../pages/provider';
-import { runTestAs } from '../../../helpers';
+import { AllClaimsPage } from '../pages/caseworker/all-claims';
+import {
+    authenticateAsCaseworker,
+    formData,
+    formatDate,
+    caseworkerAppUrl,
+    authenticateAsProvider
+} from '../../../helpers';
+
 export default function createTests() {
     let page;
+    let laaReference;
     test.describe.configure({ mode: 'serial' });
 
     test.beforeAll(async ({ browser }) => {
@@ -23,11 +32,10 @@ export default function createTests() {
     test.afterAll(async () => {
         await page.close();
     });
-    // Authenticate as provider
-    test.use({ storageState: runTestAs('provider') })
 
     // Checking current risk level
     test(' Start a new claim with No Disbursements and no uploads', async () => {
+        await authenticateAsProvider(page);
         await test.step('Starting a new claim', async () => {
             const yourClaimsPage = new YourClaimsPage(page);
             // Actions
@@ -45,6 +53,10 @@ export default function createTests() {
         await whatAreYouClaimingPage.fillClaimForm();
         // Your claim progress
         await expect(page.getByRole('heading', { name: 'Your claim progress' })).toBeVisible();
+
+        const asideLocator = await page.locator('.aside-task-list');
+        const asideText = await asideLocator.textContent();
+        laaReference = asideText.split('LAA reference')[1].trim();
     });
 
     test('Go to details and select Your details', async () => {
@@ -165,6 +177,71 @@ export default function createTests() {
         await expect(page.getByRole('heading', { name: 'Confirm the following' })).toBeVisible();
         await page.getByLabel('Full name').fill('Test Automate');
         await page.getByRole('button', { name: 'Save and submit' }).click();
+    });
 
+    test('View claim in caseworker app', async () => {
+        const syncUrl = caseworkerAppUrl() + '/sync';
+        await page.goto(syncUrl);
+
+        await authenticateAsCaseworker(page);
+
+        const allClaimsPage = new AllClaimsPage(page);
+        await allClaimsPage.goto();
+        await expect(page.locator('#main-content')).toContainText(laaReference);
+        await page.getByRole('link', { name: laaReference }).click();
+        await page.waitForURL('**/claim_details');
+        await expect(page.getByRole('heading', { name: `${formData.defendant.firstName} ${formData.defendant.lastName}` })).toBeVisible();
+    });
+
+    test('Verify claim details', async () => {
+        const texts = [
+            `Unique file number${formData.uniqueFile}`,
+            `Type of claim${formData.claimType}`,
+            `Representation order date${formatDate(formData.repOrderDate)}`,
+            `Main defendant full name${formData.defendant.firstName} ${formData.defendant.lastName}`,
+            `Main defendant MAAT ID${formData.defendant.maatId}`,
+            'Main offence nameCriminal Damage',
+            `Offence date${formatDate(formData.mainOffenceDate)}`,
+            'Assigned counselNo',
+            'Unassigned counselNo',
+            'Instructed agentNo',
+            "Case remitted from Crown Court to magistrates' courtNo",
+            'Category 1Guilty plea',
+            "Why are you claiming a non-standard magistrates' payment?Enhanced rates claimed",
+            `Number of pages of prosecution evidence${formData.evidencePages.prosecution}`,
+            `Number of pages of defence statements${formData.evidencePages.defence}`,
+            `Number of witnesses${formData.witnesses}`,
+            'Supplemental claimNo',
+            'Recorded evidenceNo',
+            'Work done before order was grantedNo',
+            'Work was done after last hearingNo',
+            `Date of first hearing${formatDate(formData.hearingDate)}`,
+            `Number of hearings${formData.hearingCount}`,
+            "Magistrates' courtCity of London Magistrates' Court",
+            'Court is in designated area of the firmNo',
+            'Youth courtNo',
+            `Hearing outcome${formData.hearingOutcome}`,
+            `Matter type${formData.matterType}`,
+            'Any other informationNo',
+            'Proceedings concluded over 3 months agoNo',
+            `Firm name${formData.firmName}`,
+            `Firm account number${formData.firmAccountNumber}`,
+            `Firm address${formData.addressLine1}${formData.townOrCity}${formData.postcode}`,
+            `Solicitor full name${formData.solicitorFirstName} ${formData.solicitorLastName}`,
+            `Solicitor reference number${formData.solicitorReferenceNumber}`,
+            'Alternative contact detailsNot provided',
+            'Provider email addresssingle@office.com',
+            'Equality questionsNo, skip the equality questions'
+        ]
+
+        texts.forEach(async (text) => {
+            await expect(page.locator('#main-content')).toContainText(text);
+        });
+    })
+
+    test('Verify claim is for correct amount', async () => {
+        page.getByRole('link', { name: 'Adjustments' }).click();
+        await page.waitForURL('**/adjustments');
+        await expect(page.locator('#main-content')).toContainText('£624.81');
     });
 }
